requirements.txt
fastapi
python-dotenv
pinecone-client
sentence-transformers
beautifulsoup4
langchain
uvicorn
